# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VetManager Pro API
  description: |
    API REST para el sistema veterinario SaaS multinacional VetManager Pro.

    **Autenticación:** Bearer JWT (access token de 15min + refresh token de 7 días)

    **Multi-tenant:** Todos los endpoints filtran datos por `branchId` del token JWT.

    **RBAC:** Cada endpoint requiere un permiso específico según el rol del usuario.
  version: "1.0.0"
  contact:
    name: VetManager Pro
    email: api@vetmanagerpro.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.vetmanagerpro.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      summary: Listar turnos
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
        - in: query
          name: vetId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: patientId
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Lista de turnos }
    post:
      tags: [Appointments]
      summary: Crear turno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId, vetId, scheduledDate]
              properties:
                patientId:         { type: integer }
                vetId:             { type: integer }
                scheduledDate:     { type: string, format: date-time }
                durationMinutes:   { type: integer, default: 30 }
                appointmentTypeId: { type: integer }
                reason:            { type: string }
                isEmergency:       { type: boolean }
      responses:
        '201': { description: Turno creado }
...
  /appointments/{id}/status:
    patch:
      tags: [Appointments]
      summary: Cambiar estado del turno
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [scheduled,confirmed,in_progress,completed,cancelled,no_show] }
      responses:
        '204': { description: Estado actualizado }
  /appointments/{id}/triage:
    post:
      tags: [Appointments]
      summary: Registrar triage de emergencia
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority, chiefComplaint]
              properties:
                priority:       { type: string, enum: [immediate,urgent,less_urgent,non_urgent] }
                chiefComplaint: { type: string }
                heartRate:      { type: number }
                temperature:    { type: number }
                painScore:      { type: integer, minimum: 0, maximum: 10 }
      responses:
        '201': { description: Triage registrado }

  # ── MEDICAL RECORDS ────────────────────────────────────────────────────────
...
  /grooming/appointments:
    get:
      tags: [Grooming]
      summary: Listar turnos de grooming
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
        - in: query
          name: groomerId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Lista de turnos }
    post:
      tags: [Grooming]
      summary: Crear turno de grooming
      description: Calcula precio estimado automáticamente basado en servicios y tamaño del paciente.
      responses:
        '201': { description: Turno creado con precio estimado }
...
  /reports/appointments:
    get:
      tags: [Reports]
      summary: Funnel de turnos por período
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: Resumen y distribución por tipo de turno }
...